bsjktest(splm)	R Documentation

Baltagi Song Jung and Koh LM test for spatial panels

Description

Baltagi, Song, Jung and Koh marginal or conditional LM test for spatial error correlation, serial error correlation or random effects in panel models

Usage


bsjktest(formula, data, listw, ...)

## S3 method for class 'formula':
bsjktest(formula, data, listw, test=c("C.1","C.2", "C.3","J"))

Arguments

formula 	 A model formula

data 	 A data frame

listw 	 An object of either class listw or class matrix

Details

If test=J the joint test for spatial or serial error correlation or random individual effects is returned; if test is one of C.1, C.2 or C.3 the conditional test for, respectively, spatial errror correlation; serial error correlation; random individual effects; allowing for the other two effects to be present is returned.

Value

An object of class htest

Author(s)

Giovanni Millo

References

Baltagi, B.H., Song, S.H., Jung xx and Koh, W. (2007) Testing panel data regression models with spatial and serial error correlation. Journal of Econometrics, xxx, xxx–xxx.

See Also

bsktest

Examples


data(Produc, package="Ecdat")
Produc <- Produc[Produc$year<1975, ]
data(usaww)
fm <- log(gsp)~log(pcap)+log(pc)+log(emp)+unemp
test1<-bsjktest(fm,data=Produc, w=usaww,
  test="C.1")
test1

[Package splm version 0.1-1 Index]