bsktest(splm)	R Documentation

Baltagi Song and Koh LM test for spatial panels

Description

Baltagi, Song and Koh marginal or conditional LM test for spatial error correlation or random effects in panel models

Usage


bsktest(x, w, ...)
## S3 method for class 'formula':
bsktest(x, data, w, test=c("SLM1","SLM2","LMJOINT","CLMlambda","CLMmu"), index=NULL, ...)
## S3 method for class 'lm':
bsktest(x, w, index, test=c("SLM1","SLM2","LMJOINT"), ...)
## S3 method for class 'splm':
bsktest(x, w, index, test=c("CLMlambda","CLMmu"), ...)

Arguments

x 	 an object whose class is one of formula or lm or splm

data 	 an object of class data.frame. An optional data frame containing the variables in the model.The first two columns may contain the indexes. See index

index 	 if not NULL (default), a vector containing the individual and time indexes

w 	 an object of class listw created for example by nb2listw

test 	 one of "SLM1","SLM2","LMJOINT","CLMlambda","CLMmu", the test to be performed

Value

an object of class 'htest'

Author(s)

Gianfranco Piras

References

Baltagi, B.H., Song, S.H. and Koh, W. (2003) Testing panel data regression models with spatial error correlation. Journal of Econometrics, 117, 123–150.

See Also

bsjktest

Examples


data(Produc, package="Ecdat")
Produc <- Produc[Produc$year<1975, ]
data(usaww)
fm <- log(gsp)~log(pcap)+log(pc)+log(emp)+unemp
test1<-bsktest(fm,data=Produc, w=mat2listw(usaww),
  test="SLM1")
class(test1)
test1
#ml2 <- spfeml(fm, data = Produc, , mat2listw(usaww), model = "error", effects = "pooled")
#class(ml2)
#test5bis<-bsktest(ml2, w=mat2listw(usaww),index=Produc[,c(1,2)] ,test="CLMmu")
#summary(test5bis)

[Package splm version 0.1-1 Index]