effects.splm(splm)	R Documentation

method for extracting fixed effects

Description

Methods used for extracting fixed effects from objects of class 'splm' where type is one of "fixed effects lag" or "fixed effects error"

Usage


effects.splm(x)


Arguments

x 	 an object of class 'splm'

Details

If the argument x is not of class 'splm' the function will terminate with an error message.

If the argument x is of class 'splm' type is not one of "fixed effects lag" or "fixed effects error", the function will terminate with an error message.

Value

An object of class "effects.splm"

res 	 a list whose elements are various type of fixed effects and the intercept (when present)

Author(s)

Gianfranco Piras gpiras@mac.com

References

Elhorst, J.P. (2003) Specification and estimation of spatial panel data models, International Regional Science Review, 26, pages 244–268.

Elhorst, J.P. (2009) Spatial panel data models, In Fischer, M.M. and Getis, A. (eds), Handbook of Applied Spatial Analysis Springer, Berlin.

See Also

spfeml print.effects.splm summary.effects.splm

Examples


data(Produc, package = "Ecdat")
data(usaww)
Produc <- Produc[Produc$year<1975, ]
fm <- log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp
lag <- spfeml(fm, data = Produc, listw = mat2listw(usaww), effects = "sptpfe", method = "eigen", quiet = TRUE) 
summary(lag)
eff <- effects(lag) 
err <- spfeml(fm, data = Produc, listw = mat2listw(usaww), model = "error", effects = "tpfe", method = "eigen", quiet = TRUE)
summary(err)
eff <- effects(err) 
summary(eff)

[Package splm version 0.1-1 Index]