spfeml(splm)	R Documentation

Spatial Panel Fixed Effects Models Estimation

Description

Maximum likelihood (ML) estimation of spatial panel fixed effects models

Usage


spfeml(formula, data, index=NULL,listw, model=c("lag","error"),
effects=c('pooled','spfe','tpfe','sptpfe'), method="eigen",
na.action=na.fail,quiet=TRUE,zero.policy = FALSE,
 interval = c(-1, 0.999), tol.solve = 1e-10, tol.opt = .Machine$double.eps^0.5)
## S3 method for class 'spfeml':
print(object, ...)
## S3 method for class 'spfeml':
summary(object, ...)

Arguments

formula 	 a description of the model to be fit

data 	 an object of class data.frame. An optional data frame containing the variables in the model. The first two columns may contain the indexes. see index

index 	 if not NULL (default), a vector containing the individual and time indexes

listw 	 an object of class listw created for example by nb2listw

model 	 may be set to "lag" or "error"

effects 	 one of "pooled" (no spatial effects, i.e. OLS on the pooled model), "spfe" (only spatial fixed effects), "tpfe" (only time period fixed effects), "sptpfe" (both time period and spatial fixed effects)

method 	 "eigen" (default) - the Jacobian is computed as prod_{i=1}^N (1 - rhoomega_i) (with omega_i the eigenvalues of the spatial weigths matrix) using eigenw, and "spam" or "Matrix" using code from the spam or Matrix packages to calculate the determinant (see lagsarlm for details)

na.action 	 a function (default na.fail), can also be na.omit or na.exclude with consequences for residuals and fitted values - (see lagsarlm for details)

zero.policy 	 if TRUE assign zero to the lagged value of zones without neighbours, if FALSE (default) assign NA - causing spfeml to terminate with an error

interval 	 search interval for spatial parameter when not using method="eigen"; default is c(-1,1); method="Matrix" will attempt to search for an appropriate interval

tol.solve 	 tolerance for detecting linear dependence in the columns of matrices to be inverted (default= 1.0e-10)- (see lagsarlm for details)

tol.opt 	 desired accuracy for optimization (default=square root of double precision machine tolerance)

Details

The function deals with both fixed effects spatial lag and error models, of the form:

y = λ (I_T otimes W_N)y + (iota_T otimes α) + Xβ + ε

where λ is called the spatial autoregressive coefficient, and:

y = (iota_T otimes α) + Xβ + u

u = rho (iota_T otimes W_N)u + ε

respectively, where rho is the spatial autocorrelation coefficient. Both λ and rho are found by optimize, and the model parameters by a feasible GLS procedure.

The asymptotic standard error of λ is only computed when method=eigen, because the full matrix operations involved would be costly for dimensions associated with the choice of "spam" or "Matrix". The same applies to the coefficient covariance matrix.

The model variables are first transformed depending on the typology of fixed effects. As an example, if time period fixed effects are considered, the demeaned form is obtained by subtracting the average for each cross-sectional unit computed over the time dimension.

The transformed variables are then used in a concentrated likelihood approach to estimate the spatial parameters.

Once the spatial parameters are obtained, the model parameters are then estimated by a feasible GLS procedure.

Fixed effects can then be extracted using the function effects.

Value

An object of class "splm".

coefficients 	 coefficients estimate of the model parameters

vcov 	 the asymptotic variance covariance matrix of the estimated coefficients (when calculated)

type 	 'fixed effects spatial lag (or error) model'

spat.coeff 	 the spatial coefficients rho or λ

residuals 	 the GLS residuals

fitted.values 	 difference between response variable and residuals

sigma2 	 GLS residuals variance

model 	 the matrix of the data used

call 	 the call used to create the object

logLik 	 the value of the log likelihood function at the optimum

method 	 the method used to calculate the Jacobian

effects 	 the typology of effects included in the model

res.eff 	 a list whose elements are the fixed effects (see effects)

Author(s)

Gianfranco Piras gpiras@mac.com

References

Elhorst, J.P. (2003) Specification and estimation of spatial panel data models, International Regional Science Review, 26, pages 244–268.

Elhorst, J.P. (2009) Spatial panel data models, In Fischer, M.M. and Getis, A. (eds), Handbook of Applied Spatial Analysis Springer, Berlin.

Anselin, L. and Le Gallo, J. and Jayet, H. (2008) Spatial Panel Econometrics, In Matyas, L. and Sevestre, P. (eds), The econometrics of Panel Data, Fundamentals and Recent Developments in Theory and Practice (3rd Edition), pages 624–660. Springer-Verlag, Berlin Heidelberg.

See Also

spreml, spregm, lagsarlm, errorsarlm, effects

Examples


data(Produc, package = "Ecdat")
data(usaww)
Produc <- Produc[Produc$year<1975, ]
fm <- log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp
lag <- spfeml(fm, data = Produc, listw = mat2listw(usaww), effects = "sptpfe", method = "eigen", quiet = TRUE)
summary(lag)
eff <- effects(lag)
err <- spfeml(fm, data = Produc, listw = mat2listw(usaww), model = "error", effects = "tpfe", method = "eigen", quiet = TRUE)
summary(err)
eff <- effects(err)
summary(eff)

[Package splm version 0.1-1 Index]