splm-package(splm)	R Documentation

Spatial panel models: estimation and testing

Description

A comprehensive toolset for ML and GM estimation and diagnostic testing of econometric models for spatial panel data.

Details

Package: 	 splm

Type: 	 Package

Version: 	 0.0

Date: 	 2009-07-29

License: 	 GPL

LazyLoad: 	 yes

Put here: An overview of how to use the package, including the most important functions.

Author(s)

Giovanni Millo and Gianfranco Piras

Maintainer: Giovanni Millo <giovanni.millo@generali.com>

References

Here: Literature or other references for background information

Examples


##Here put simple examples of the most important functions

[Package splm version 0.1-1 Index]