spregm(splm)	R Documentation

GM estimator for spatial panel data models

Description

GM estimator for panel data models with spatially correlated errors components of the form:

y_N(t) = X_N(t) β + u_N(t)

u_N(t) = rho W_N u_N(t) + ε(t)

ε_N = (e_T otimes I_N ) μ_N + nu_N

where rho, and the variance components σ^2_μ and σ^2_nu are estimated by GM, and the model coefficients by GLS.

Usage


spregm(formula, data=list(), index=NULL, w, method = c("init", "weigh", "fulweigh"))
## S3 method for class 'spregm':
print(object, ...)
## S3 method for class 'spregm':
summary(object, ...)

Arguments

formula 	 a description of the model to be fit. The details of model specification are given for lm

data 	 an object of class data.frame. An optional data frame containing the variables in the model. The first two columns may contain the indexes. See index

index 	 if not NULL (default), a vector containing the individual and time indexes

w 	 an object of class listw created for example by nb2listw. If a matrix is given as input the function will transform it using mat2listw

method 	 "init" (default) defines the set of GM estimator to be used. Alternatives are "weigh" and "fulweigh" (See Details)

Details

When method=init, the initial estimator is calculated. This first set of GM estimators is based only on a subset of the moments conditions and assigns equal weigths to each of them. When method=fulweigh, the second set of GM estimators is calculated. This estimator is based on the full set of moments conditions. It also involves the expression for the variance covariance matrix of the sample moments calculated under the assumption of normally distributed innovations. The calculation of the trace terms in the expression of the variance covariance matrix of the sample moments uses codes from the Matrix package. When method=weigh, the third set of GM estimator is used. This is motivated by computational issues. The procedure is analogous to the second one but uses a simplified expression for the variance covariance matrix of the sample moments

Note that σ^2_μ is not reported. σ^2_1 is reported instead. However, a value for σ^2_μ can easily be obtained from:

σ^2_1 = σ^2_nu + T σ^2_μ

The function also produces an estimate for theta which is a function of the variance components.

Value

An object of class "splm".

coefficients 	 GLS coefficients estimate of the model parameters

vcov 	 the variance covariance matrix of the estimated coefficients

residuals 	 the GLS residuals

fitted.values 	 difference between response variable and residuals

sigma2 	 GLS residuals variance

type 	 'random effect GM'

rho 	 a vector including the spatial parameter and the variance components (see Details)

model 	 the matrix of the data used

call 	 the call used to create the object

Author(s)

Gianfranco Piras gpiras@mac.com

References

Kapoor, M., Kelejian, H.H. and Prucha, I.R. (2007) Panel data model with spatially correlated error components, Journal of Econometrics, 140, pages 97–130.

Kelejian, H.H. and Prucha, I.R. (1999) A Generalized Moments Estimator for the Autoregressive Parameter in a Spatial Model, International Economic Review, 40, pages 509–533.

Kelejian, H.H. and Prucha, I.R. (1999) A Generalized Spatial Two Stage Least Square Procedure for Estimating a Spatial Autoregressive Model with Autoregressive Disturbances, Journal of Real Estate Finance and Economics, 17, pages 99–121.

See Also

spreml, spsegm

Examples


data(Produc, package = "Ecdat") 
data(usaww)
Produc <- Produc[Produc$year<1975, ] 
GM<-spregm(log(gsp)~log(pcap)+log(pc)+log(emp)+unemp, data=Produc,w=usaww,method="fulweigh")
summary(GM)

[Package splm version 0.1-1 Index]