spreml(splm)	R Documentation

Spatial Panel Random Effects Model Estimation

Description

Maximum likelihood (ML) estimation of spatial random effects panel

Usage


spreml(formula, data, w, subset = NULL, na.action = "na.fail",
    effect = "individual", model = "random", lag = FALSE,
    errors = c("semsrre", "semsr", "ssrre", "semre", "re", "ssr",
"sem"),
    index = NULL, pvar = FALSE, quiet = TRUE, hess = FALSE,
    tol = 0.01, ltol = 1000, starting.values = "pooling", ...)
## S3 method for class 'spreml':
print(object, ...)
## S3 method for class 'spreml':
summary(object, ...)

Arguments

formula 	 a symbolic description of the model to be estimated

data 	 a data.frame

w 	 an object of either class listw or class matrix

subset 	 see lm

na.action 	 see lm

effect 	 for compatibility with plm; this must always be ''individual''

model 	 for compatibility with plm; this must always be ''random''

lag 	 if TRUE include a lagged dependent variable

errors 	 covariance structure of the errors: one of "semsr", "ssrre", "semre", "re", "ssr", "sem"

index 	 if not NULL (default), a vector containing the individual and time indexes

pvar 	 if TRUE, the pvar function is called

quiet 	 if FALSE the parameter values are printed at each (major) iteration

hess 	 if TRUE use numerical Hessian instead of GLS for the standard errors of the estimates

tol 	 tolerance for the first-step optimization

ltol 	 tolerance for the internal step, to be passed on to the factr argument of optim

starting.values 	 vector of starting values for the estimates: one of ''pooling'', ''random'' or ''within'' (see plm)

... 	 further arguments to be passed on to the optimizer

Details

The models are estimated by two-step Maximum Likelihood.

The covariance structures allowed are:
semsr

spatial and serial correlation in the idiosyncratic error term

ssrre

individual random effects and serial correlation

semre

individual random effects and spatial correlation

re

individual random effects

ssr

serial correlation

sem

spatial correlation

Value

An object of class "splm".

coefficients 	 coefficients estimate of the model parameters

arcoef 	 the coefficient for the spatial lag on y

errcomp 	 the estimates of the error variance components

vcov 	 the asymptotic variance covariance matrix of the estimated coefficients

vcov.arcoef 	 the asymptotic variance of the estimated spatial lag parameter

vcov.errcomp 	 the asymptotic variance covariance matrix of the estimated error covariance parameters

type 	 'random effects ML'

residuals 	 the model residuals

fitted.values 	 the fitted values, calculated as \hat{y}=X \hat{\beta}

sigma2 	 GLS residuals variance

model 	 the matrix of the data used

call 	 the call used to create the object

logLik 	 the value of the log likelihood function at the optimum

errors 	 the value of the errors argument

Author(s)

Giovanni Millo

See Also

spregm, spfeml

Examples


data(Produc, package = "Ecdat")
data(usaww)
Produc <- Produc[Produc$year<1974, ]
fm <- log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp
respaterr <- spreml(fm, data = Produc, w = usaww, errors="semre")
summary(respaterr)

[Package splm version 0.1-1 Index]