summary.splm(splm)	R Documentation

summary method for class splm

Description

Methods used for presenting the results of objects of class 'splm'

Usage



## S3 method for class 'splm':
summary(object,...)

Arguments

object 	 an object of class 'splm'

Details

The summary function summary.splm returns an objects of class 'splm' organized in a coefficient matrix.

Also a matrix for the error components, or the spatial coefficient will be generated depending on the estimated model.

When the 'splm' objects comes from the function 'spsegm', the summary will be generated looping over the number of equations in the system.

Value

An object of class 'summary.splm'

Author(s)

Giovanni Millo Giovanni_Millo@Generali.com, Gianfranco Pirasgpiras@mac.com

See Also

spreml, spregm, lagsarlm, errorsarlm, spreml, effects

Examples


data(Produc, package = "Ecdat") 
data(usaww)
Produc <- Produc[Produc$year<1975, ] 
GM<-spregm(log(gsp)~log(pcap)+log(pc)+log(emp)+unemp, data=Produc,w=usaww,method="fulweigh")
summary(GM)

[Package splm version 0.1-1 Index]